+++ realm/realm-library/src/syncIntegrationTest/kotlin/io/realm/SyncSessionTests.kt
@@ -171,7 +171,7 @@ class SyncSessionTests {
         val config = configFactory.createSyncConfigurationBuilder(user, "e873fb25-11ef-498f-9782-3c8e1cd2a12c")
             .assetFile("synced_realm_e873fb25-11ef-498f-9782-3c8e1cd2a12c_no_client_id.realm")
             .syncClientResetStrategy(object:
-                AutomaticRecoverUnsyncedChangesStrategy {
+                RecoverUnsyncedChangesStrategy {
                 override fun onError(session: SyncSession, error: ClientResetRequiredError) {
                     fail("This test case was not supposed to trigger AutomaticRecoveryStrategy::onError()")
                 }
@@ -191,7 +191,7 @@ class SyncSessionTests {
         val config = configFactory.createSyncConfigurationBuilder(user, "e873fb25-11ef-498f-9782-3c8e1cd2a12c")
             .assetFile("synced_realm_e873fb25-11ef-498f-9782-3c8e1cd2a12c_no_client_id.realm")
             .syncClientResetStrategy(object:
-                AutomaticRecoverOrDiscardUnsyncedChangesStrategy {
+                RecoverOrDiscardUnsyncedChangesStrategy {
                 override fun onBeforeReset(realm: Realm) {
                     fail("This test case was not supposed to trigger AutomaticRecoveryOrDiscardUnsyncedChangesStrategy::onBeforeReset()")
                 }
@@ -225,7 +225,7 @@ class SyncSessionTests {
         }
         val config = configFactory.createSyncConfigurationBuilder(user, "e873fb25-11ef-498f-9782-3c8e1cd2a12c")
             .syncClientResetStrategy(object:
-                AutomaticRecoverOrDiscardUnsyncedChangesStrategy {
+                RecoverOrDiscardUnsyncedChangesStrategy {
                 override fun onBeforeReset(realm: Realm) {
                     assertTrue(realm.isFrozen)
                     Assert.assertEquals(1, realm.where<SyncColor>().count())